Tests/LibWeb/Text/expected/HTML/button-layout-image-percentage-height.txt

button 1207x32
image 229x32
